pyproject.toml
[project]
name = "strategic-analyst-agent"
version = "0.1.0"
description = "Personalised ambient strategic analyst agent built on LangGraph"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # LangGraph core
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.4.0",
    "langgraph-checkpoint-postgres>=3.0.4",

    # LangChain ecosystem
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langsmith>=0.2.0",

    # Supabase
    "supabase>=2.4.0",
    "vecs>=0.4.0",
    "psycopg[binary]>=3.1.0",

    # Web search
    "tavily-python>=0.3.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "aiohttp>=3.9.0",

    # Report generation
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-env>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/strategic_analyst"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
